fix: Fix typo in comment.model.js and cloudinary.js
This commit fixes a typo in the `comment.model.js` file and the `cloudinary.js` file. In the `comment.model.js` file, the export name was misspelled as "Commnet" instead of "Comment". This commit corrects the export name to "Comment". In the `cloudinary.js` file, a new function `deleteOnCloudinary` was added, but it was not exported. This commit adds the export statement for the `deleteOnCloudinary` function.